# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Corporate Gifting ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/vendors:
    get:
      operationId: listVendors
      tags: [purchase-orders]
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [purchase-orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /v1/vendors/{id}:
    get:
      operationId: getVendor
      tags: [purchase-orders]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [purchase-orders]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # ── Sales Orders ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        contactPerson: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        leadTimeDays: { type: integer }
        createdAt: { type: string }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: string }
        leadTimeDays: { type: integer }
    VendorUpdate:
      type: object
      properties:
        name: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: string }
        leadTimeDays: { type: integer }

    # ── Sales Orders ──
...